Fix formating and other issues
Refactor as per comments.

Simplified some of the code as well

Use constants

+++ src/openrct2/ReplayManager.cpp
@@ -506,12 +506,12 @@ namespace OpenRCT2
                 }
                 case GAME_COMMAND_PLACE_RIDE_ENTRANCE_OR_EXIT:
                 {
-                    CoordsXY loc = {command.eax & 0xFFFF, command.ecx & 0xFFFF};
+                    CoordsXY loc = { (int32_t)(command.eax & 0xFFFF), (int32_t)(command.ecx & 0xFFFF) };
                     Direction direction = (command.ebx >> 8) & 0xFF;
                     ride_id_t rideId = command.edx & 0xFF;
                     uint8_t stationNum = command.edi & 0xFF;
                     bool isExit = ((command.edx >> 8) & 0xFF) != 0;
-                    result.action = std::make_unique<RideEntranceExitPlaceAction>(loc,direction,rideId,stationNum,isExit);
+                    result.action = std::make_unique<RideEntranceExitPlaceAction>(loc, direction, rideId, stationNum, isExit);
                     result.action->SetFlags(command.ebx & 0xFF);
                     break;
                 }
